+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/State/Query.hs
@@ -652,17 +652,16 @@ queryStakeSnapshots ::
   Maybe (Set (KeyHash StakePool)) ->
   StakeSnapshots
 queryStakeSnapshots nes mPoolIds =
-  let SnapShots
-        { ssStakeMark
-        , ssStakeSet
-        , ssStakeGo
-        } = esSnapshots $ nesEs nes
+  let snaps = esSnapshots $ nesEs nes
+      markSnap = snaps ^. ssStakeMarkL . msSnapShotL
+      setSnap = snaps ^. ssStakeSetL . ssSnapShotL
+      goSnap = snaps ^. ssStakeGoL . gsSnapShotL
 
       mkStakeSnapshotMaybe poolId = do
         let
-          markPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeMark)
-          setPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeSet)
-          goPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot ssStakeGo)
+          markPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot markSnap)
+          setPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot setSnap)
+          goPoolStake = spssStake <$> VMap.lookup poolId (ssStakePoolsSnapShot goSnap)
         -- Non-registered stake pools or ones that have no stake are of no interest to us.
         guard (fold [markPoolStake, setPoolStake, goPoolStake] > Just mempty)
         Just
@@ -677,9 +676,9 @@ queryStakeSnapshots nes mPoolIds =
             maybe mempty (fromCompact . spssStake) . VMap.lookup poolId . ssStakePoolsSnapShot
          in
           StakeSnapshot
-            { ssMarkPool = lookupStake ssStakeMark
-            , ssSetPool = lookupStake ssStakeSet
-            , ssGoPool = lookupStake ssStakeGo
+            { ssMarkPool = lookupStake markSnap
+            , ssSetPool = lookupStake setSnap
+            , ssGoPool = lookupStake goSnap
             }
       version = pvMajor (nes ^. nesEsL . curPParamsEpochStateL . ppProtocolVersionL)
       poolIds =
@@ -688,20 +687,20 @@ queryStakeSnapshots nes mPoolIds =
             | version < natVersion @11 ->
                 foldMap
                   (VMap.keysSet . VMap.filter (\_ -> (> 0) . spssNumDelegators) . ssStakePoolsSnapShot)
-                  [ssStakeMark, ssStakeSet, ssStakeGo]
+                  [markSnap, setSnap, goSnap]
             | otherwise ->
                 foldMap
                   (VMap.keysSet . VMap.filter (\_ -> (> mempty) . spssStake) . ssStakePoolsSnapShot)
-                  [ssStakeMark, ssStakeSet, ssStakeGo]
+                  [markSnap, setSnap, goSnap]
           Just ids -> ids
    in StakeSnapshots
         { ssStakeSnapshots =
             if version < natVersion @11
               then Map.fromSet mkStakeSnapshot poolIds
               else Map.mapMaybe id $ Map.fromSet mkStakeSnapshotMaybe poolIds
-        , ssMarkTotal = ssTotalActiveStake ssStakeMark
-        , ssSetTotal = ssTotalActiveStake ssStakeSet
-        , ssGoTotal = ssTotalActiveStake ssStakeGo
+        , ssMarkTotal = ssTotalActiveStake markSnap
+        , ssSetTotal = ssTotalActiveStake setSnap
+        , ssGoTotal = ssTotalActiveStake goSnap
         }
 
 -- | Query the current epoch number.