+++ libs/cardano-ledger-test/bench/Bench/Cardano/Ledger/StakeDistr.hs
@@ -239,7 +239,7 @@ tickfRuleBench =
             ]
 
 getSnap :: NewEpochState AlonzoEra -> SnapShot
-getSnap nes = ssStakeSet $ esSnapshots $ nesEs nes
+getSnap nes = ssSnapShot $ ssStakeSet $ esSnapshots $ nesEs nes
 
 getRewardUpdate :: NewEpochState era -> RewardUpdate
 getRewardUpdate nes =